Worksphere/workshpere/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.2
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
beautifulsoup4==4.12.2
python-dotenv==1.0.0
python-multipart==0.0.6